+++ doc/CHANGES.rst
@@ -59,6 +59,7 @@ Future
 * [REF] the foreign key must be declared with ModelAttribute
 * [REF] Use Adapter for Model and attribute in relation ship
 * [REF] hybrid_method, cache and classmethod_cache are now only impotable decorator function
+* [IMP] in column the default can be a classmethod name
 
 0.4.1
 -----